OWIntegrate: fix non-meta output
I created that bug when I also started processing reference.
